Remove non-generic vector table for K64F
This patch removes all but the first 16 entries from the K64F's vector
table. These are installed at runtime and do not need to be part of
the ROM vector table. This allows functions that are not used to be
removed by the linker.

Additionally this patch updates the K64F linker script to place code
where the extra vectors would have been.

This saves > 2K rom when building as development with ARMCC.